fix: move GeoJsonContinuerIsCompleted to app/bpartners/geojobs/endpoint/event/model
